updated hbm.xml files & hibernate-Session, new Inserts & Selects, SQLXML
new methods:
	Inserts.unit
	Inserts.rule
	Selects.deviceComponent
	Selects.rulesByDeviceComponentID
	Selects.rule

Hibernate doesn´t work with XML so i created SQLXMLUserType class 

Revised session.evict

Every DB-Method close the session (session.close())

+++ Boese/src/main/java/de/bo/aid/boese/model/DeviceComponentRuleId.java
@@ -8,7 +8,7 @@ public class DeviceComponentRuleId implements java.io.Serializable {
 	public DeviceComponentRuleId() {
 	}
 
-	public DeviceComponentRuleId(int deCoId, short ruId) {
+	public DeviceComponentRuleId(int deCoId, int ruId) {
 		this.deCoId = deCoId;
 		this.ruId = ruId;
 	}
@@ -25,7 +25,7 @@ public class DeviceComponentRuleId implements java.io.Serializable {
 		return this.ruId;
 	}
 
-	public void setRuId(short ruId) {
+	public void setRuId(int ruId) {
 		this.ruId = ruId;
 	}
